Simplified analysis for time rabi chevron, removed init_utils.py
